requirements.txt
playwright==1.49.1
notion-client==2.2.1
python-dotenv==1.0.1
rich==13.9.4
requests==2.32.3
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
pywebview==5.3.2
pyinstaller==6.11.1
